+++ librelane/common/metrics/library.py
@@ -124,6 +124,12 @@ Metric(
 )
 
 Metric(
+    "timing__clock__fmax",
+    aggregator=min_aggregator,
+    higher_is_better=True,
+)
+
+Metric(
     "timing__setup_r2r__ws",
     aggregator=min_aggregator,
     higher_is_better=True,